+++ packages/svelte/src/compiler/phases/2-analyze/validation.js
@@ -578,6 +578,8 @@ const validation = {
 		});
 	},
 	RenderTag(node, context) {
+		context.state.analysis.uses_render_tags = true;
+
 		const raw_args = unwrap_optional(node.expression).arguments;
 		for (const arg of raw_args) {
 			if (arg.type === 'SpreadElement') {
@@ -1172,6 +1174,27 @@ export const validation_runes = merge(validation, a11y_validators, {
 			) {
 				warn(state.analysis.warnings, node, path, 'derived-iife');
 			}
+		}
+	},
+	AssignmentPattern(node, { state, path }) {
+		if (
+			node.right.type === 'Identifier' &&
+			node.right.name === '$bindable' &&
+			!state.scope.get('bindable')
+		) {
+			warn(state.analysis.warnings, node, path, 'invalid-bindable-declaration');
+		}
+	},
+	SlotElement(node, { state, path }) {
+		if (!state.analysis.custom_element) {
+			warn(state.analysis.warnings, node, path, 'deprecated-slot-element');
+		}
+	},
+	OnDirective(node, { state, path }) {
+		const parent_type = path.at(-1)?.type;
+		// Don't warn on component events; these might not be under the author's control so the warning would be unactionable
+		if (parent_type === 'RegularElement' || parent_type === 'SvelteElement') {
+			warn(state.analysis.warnings, node, path, 'deprecated-event-handler', node.name);
 		}
 	},
 	// TODO this is a code smell. need to refactor this stuff