+++ packages/svelte/src/compiler/warnings.js
@@ -39,7 +39,9 @@ const runes = {
 	'derived-iife': () =>
 		`Use \`$derived.by(() => {...})\` instead of \`$derived((() => {...})());\``,
 	'invalid-props-declaration': () =>
-		`Component properties are declared using $props() in runes mode. Did you forget to call the function?`
+		`Component properties are declared using $props() in runes mode. Did you forget to call the function?`,
+	'invalid-bindable-declaration': () =>
+		`Bindable component properties are declared using $bindable() in runes mode. Did you forget to call the function?`
 };
 
 /** @satisfies {Warnings} */
@@ -231,7 +233,12 @@ const legacy = {
 		'All dependencies of the reactive declaration are declared in a module script and will not be reactive',
 	/** @param {string} name */
 	'unused-export-let': (name) =>
-		`Component has unused export property '${name}'. If it is for external reference only, please consider using \`export const ${name}\``
+		`Component has unused export property '${name}'. If it is for external reference only, please consider using \`export const ${name}\``,
+	'deprecated-slot-element': () =>
+		`Using <slot> to render parent content is deprecated. Use {@render ...} tags instead.`,
+	/** @param {string} name */
+	'deprecated-event-handler': (name) =>
+		`Using on:${name} to listen to the ${name} event is is deprecated. Use the event attribute on${name} instead.`
 };
 
 const block = {