+++ packages/svelte/src/internal/client/dom/blocks/each.js
@@ -24,7 +24,8 @@ import {
 } from '../../reactivity/effects.js';
 import { source, mutable_source, set } from '../../reactivity/sources.js';
 import { is_array, is_frozen } from '../../utils.js';
-import { STATE_SYMBOL } from '../../constants.js';
+import { INERT, STATE_SYMBOL } from '../../constants.js';
+import { push_template_node } from '../template.js';
 
 /**
  * The row of a keyed each block that is currently updating. We track this
@@ -70,7 +71,7 @@ function pause_effects(items, controlled_anchor, callback) {
 		parent_node.append(controlled_anchor);
 	}
 
-	run_out_transitions(transitions, () => {
+	run_out_transitions(transitions, true, () => {
 		for (var i = 0; i < length; i++) {
 			destroy_effect(items[i].e);
 		}
@@ -168,10 +169,11 @@ export function each(anchor, flags, get_collection, get_key, render_fn, fallback
 					break;
 				}
 
+				var child_open = /** @type {Comment} */ (child_anchor);
 				child_anchor = hydrate_anchor(child_anchor);
 				var value = array[i];
 				var key = get_key(value, i);
-				item = create_item(child_anchor, prev, null, value, key, i, render_fn, flags);
+				item = create_item(child_open, child_anchor, prev, null, value, key, i, render_fn, flags);
 				state.items.set(key, item);
 				child_anchor = /** @type {Comment} */ (child_anchor.nextSibling);
 
@@ -238,8 +240,8 @@ function reconcile(array, state, anchor, render_fn, flags, get_key) {
 	/** @type {import('#client').EachState | import('#client').EachItem} */
 	var prev = state;
 
-	/** @type {import('#client').EachItem[]} */
-	var to_animate = [];
+	/** @type {Set<import('#client').EachItem>} */
+	var to_animate = new Set();
 
 	/** @type {import('#client').EachItem[]} */
 	var matched = [];
@@ -267,7 +269,7 @@ function reconcile(array, state, anchor, render_fn, flags, get_key) {
 
 			if (item !== undefined) {
 				item.a?.measure();
-				to_animate.push(item);
+				to_animate.add(item);
 			}
 		}
 	}
@@ -278,8 +280,14 @@ function reconcile(array, state, anchor, render_fn, flags, get_key) {
 		item = items.get(key);
 
 		if (item === undefined) {
+			var child_open = /** @type {Text} */ (push_template_node(empty()));
+			var child_anchor = current ? current.o : anchor;
+
+			child_anchor.before(child_open);
+
 			prev = create_item(
-				current ? get_first_child(current) : anchor,
+				child_open,
+				child_anchor,
 				prev,
 				prev.next,
 				value,
@@ -302,14 +310,17 @@ function reconcile(array, state, anchor, render_fn, flags, get_key) {
 			update_item(item, value, i, flags);
 		}
 
-		resume_effect(item.e);
+		if ((item.e.f & INERT) !== 0) {
+			resume_effect(item.e);
+			to_animate.delete(item);
+		}
 
 		if (item !== current) {
 			if (seen.has(item)) {
 				if (matched.length < stashed.length) {
 					// more efficient to move later items to the front
 					var start = stashed[0];
-					var local_anchor = get_first_child(start);
+					var local_anchor = start.o;
 					var j;
 
 					prev = start.prev;
@@ -338,7 +349,7 @@ function reconcile(array, state, anchor, render_fn, flags, get_key) {
 				} else {
 					// more efficient to move earlier items to the back
 					seen.delete(item);
-					move(item, current ? get_first_child(current) : anchor);
+					move(item, current ? current.o : anchor);
 
 					link(item.prev, item.next);
 					link(item, prev.next);
@@ -401,20 +412,6 @@ function reconcile(array, state, anchor, render_fn, flags, get_key) {
 
 /**
  * @param {import('#client').EachItem} item
- * @returns {Text | Element | Comment}
- */
-function get_first_child(item) {
-	var current = item.e.dom;
-
-	if (is_array(current)) {
-		return /** @type {Text | Element | Comment} */ (current[0]);
-	}
-
-	return /** @type {Text | Element | Comment} */ (current);
-}
-
-/**
- * @param {import('#client').EachItem} item
  * @param {any} value
  * @param {number} index
  * @param {number} type
@@ -434,6 +431,7 @@ function update_item(item, value, index, type) {
 
 /**
  * @template V
+ * @param {Comment | Text} open
  * @param {Node} anchor
  * @param {import('#client').EachItem | import('#client').EachState} prev
  * @param {import('#client').EachItem | null} next
@@ -444,7 +442,7 @@ function update_item(item, value, index, type) {
  * @param {number} flags
  * @returns {import('#client').EachItem}
  */
-function create_item(anchor, prev, next, value, key, index, render_fn, flags) {
+function create_item(open, anchor, prev, next, value, key, index, render_fn, flags) {
 	var previous_each_item = current_each_item;
 
 	try {
@@ -462,6 +460,7 @@ function create_item(anchor, prev, next, value, key, index, render_fn, flags) {
 			a: null,
 			// @ts-expect-error
 			e: null,
+			o: open,
 			prev,
 			next
 		};
@@ -483,6 +482,8 @@ function create_item(anchor, prev, next, value, key, index, render_fn, flags) {
  * @param {Text | Element | Comment} anchor
  */
 function move(item, anchor) {
+	anchor.before(item.o);
+
 	var dom = item.e.dom;
 
 	if (dom !== null) {