+++ packages/svelte/src/reactivity/map.js
@@ -3,6 +3,9 @@ import { source, set } from '../internal/client/reactivity/sources.js';
 import { get } from '../internal/client/runtime.js';
 import { UNINITIALIZED } from '../constants.js';
 import { map } from './utils.js';
+import { INSPECT_SYMBOL } from '../internal/client/constants.js';
+
+var inited = false;
 
 /**
  * @template K
@@ -20,6 +23,22 @@ export class ReactiveMap extends Map {
 	constructor(value) {
 		super();
 
+		if (DEV) {
+			if (!inited) {
+				inited = true;
+				// @ts-ignore
+				ReactiveMap.prototype[INSPECT_SYMBOL] = function () {
+					// changes could either introduced by
+					// - modifying the value, or
+					// - add / remove entries to the map
+					for (const [, source] of this.#sources) {
+						get(source);
+					}
+					get(this.#size);
+				};
+			}
+		}
+
 		// If the value is invalid then the native exception will fire here
 		if (DEV) new Map(value);
 
@@ -28,7 +47,6 @@ export class ReactiveMap extends Map {
 
 			for (var [key, v] of value) {
 				sources.set(key, source(v));
-				super.set(key, v);
 			}
 
 			this.#size.v = sources.size;
@@ -62,7 +80,8 @@ export class ReactiveMap extends Map {
 	forEach(callbackfn, this_arg) {
 		get(this.#version);
 
-		return super.forEach(callbackfn, this_arg);
+		var bound_callbackfn = callbackfn.bind(this_arg);
+		this.#sources.forEach((s, key) => bound_callbackfn(s.v, key, this));
 	}
 
 	/** @param {K} key */
@@ -96,7 +115,7 @@ export class ReactiveMap extends Map {
 			set(s, value);
 		}
 
-		return super.set(key, value);
+		return this;
 	}
 
 	/** @param {K} key */
@@ -105,13 +124,14 @@ export class ReactiveMap extends Map {
 		var s = sources.get(key);
 
 		if (s !== undefined) {
-			sources.delete(key);
+			var removed = sources.delete(key);
 			set(this.#size, sources.size);
 			set(s, /** @type {V} */ (UNINITIALIZED));
 			this.#increment_version();
+			return removed;
 		}
 
-		return super.delete(key);
+		return false;
 	}
 
 	clear() {
@@ -126,7 +146,6 @@ export class ReactiveMap extends Map {
 		}
 
 		sources.clear();
-		super.clear();
 	}
 
 	keys() {